hterm: Make connection dialog resize with window.
This makes the connect dialog size a function of the window size.
Previously it was a hardcoded size, which made it very awkward for
nonstandard window sizes.

This change also makes it possible for the contents of the
"dialog-billboard" (the container for the form elements) to grow
scrollbars if necessary.

It also flips the initialization around, so that the connect dialog
loads without running any code.  Instead, it waits for the 'ipc-init'
message from hterm, and then does its thing.  This avoids a race
condition that caused an unusable connect dialog in some cases.

BUG=none
TEST=nassh.test(), 66/66 tests passed.

Change-Id: I70aad260a7f5ff612cdd3d839d49017c291f7e26
Reviewed-on: https://gerrit.chromium.org/gerrit/30930
Commit-Queue: Robert Ginda <rginda@chromium.org>
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>