tree_status: add option for icon clicking behavior
The code is designed to set the focus of an existing tab if one is
opened to the waterfall.  Personally, I prefer it if the button just
always opened a new tab.  Add a config option to control that.

We also tweak the options style sheet slightly to make it adding
additional sections look cleaner.  No need for all that whitespace.

BUG=None
TEST=clicking the button always creates a new tab

Change-Id: Ie74631279b7a5238fc6fabf788c9a78053f8c40f
Reviewed-on: https://gerrit.chromium.org/gerrit/44495
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>