Merging in two fixes from ChromeVox 26.
• Correctly consider SCRIPT and NOSCRIPT nodes in TraverseUtil/TraverseContent.

This is a bug where a user will cross into the text content of a SCRIPT or NOSCRIPT when navigating by character, word, or line.

• Correct a bug in the collapsing of a CursorSelection:

We collapse a selection to the start of the selection for certain commands. With the ability to move by line, we were not computing the right start offset for a line (which could occur in the middle of a node's text content).

TEST=Full test pass using TOT imaged Samsung Chromebooks with this change patched.

BUG=chromium:217658

Change-Id: Ifaa8f5012c95d4f9c5197f9927831b3fdfcf7598
Reviewed-on: https://gerrit.chromium.org/gerrit/45757
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Tested-by: David Tseng <dtseng@chromium.org>